Lançando exceções customizadas

+++ sgp-api/src/main/java/br/com/treina/recife/sgp/api/service/UsuarioService.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Service;
 
 import br.com.treina.recife.sgp.api.dto.DadosUsuarioDTO;
 import br.com.treina.recife.sgp.api.dto.UsuarioDTO;
+import br.com.treina.recife.sgp.api.exception.RecursoNaoEncontradoException;
+import br.com.treina.recife.sgp.api.exception.RegraDeNegocioException;
 import br.com.treina.recife.sgp.api.model.Usuario;
 import br.com.treina.recife.sgp.api.repository.UsuarioRepository;
 
@@ -33,51 +35,59 @@ public class UsuarioService {
 
     // SELECT * FROM TB_USUARIOS WHERE ID = ?
     public UsuarioDTO obterDadosDoUsuario(Long id) {
-        Optional<Usuario> usuario = usuarioRepository.findById(id);
-        
-        if (usuario.isPresent()) {
-            return usuario.get().toDTO();
-        }
+        Usuario usuario = usuarioRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Usuário não encontrado com o ID: " + id));
 
-        return null;
+        return usuario.toDTO();
     }
 
     // INSERT INTO TB_USUARIOS ...
     public Usuario cadastrarUsuario(DadosUsuarioDTO usuario) {
+        String cpf = usuario.cpf();
+
+        Optional<Usuario> usuarioOpt = usuarioRepository.findByCpf(cpf);
+
+        if (usuarioOpt.isPresent()) {
+            throw new RegraDeNegocioException("Já existe um usuário cadastrado com o CPF: " + cpf);
+        }
+
         return usuarioRepository.save(usuario.toModel());
     }
 
     // UPDATE TB_USUARIOS ... WHERE ID = ?
     public Usuario atualizarUsuario(Long id, DadosUsuarioDTO dados) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Não é possível atualizar. Usuário não encontrado com ID: " + id);
+        }
+
         Usuario usuario = dados.toModel();
-        usuario.setId(id);  // Garante que o usuário vai ser atualizado e não criado um novo!
+        usuario.setId(id); // Garante que o usuário vai ser atualizado e não criado um novo!
         return usuarioRepository.save(usuario);
     }
 
     // DELETE FROM TB_USUARIOS WHERE ID = ?
     public void excluirUsuario(Long id) {
+        if (!usuarioRepository.existsById(id)) {
+            throw new RecursoNaoEncontradoException("Não é possível excluir. Usuário não encontrado com ID: " + id);
+        }
+
         usuarioRepository.deleteById(id);
     }
-    
+
     // SELECT * FROM TB_USUARIOS WHERE CPF = ?
     public UsuarioDTO buscarUsuarioPeloCPF(String cpf) {
-        Optional<Usuario> usuario = usuarioRepository.findByCpf(cpf);
+        Usuario usuario = usuarioRepository.findByCpf(cpf)
+                .orElseThrow(() -> new RecursoNaoEncontradoException("Usuário não encontrado com o CPF: " + cpf));
 
-        if (usuario.isPresent()) {
-            return usuario.get().toDTO();
-        }
-
-        return null;
+        return usuario.toDTO();
     }
 
+    // SELECT * FROM TB_USUARIOS WHERE EMAIL = ? AND SENHA = ?
     public UsuarioDTO buscarUsuarioPeloEmailSenha(String email, String senha) {
-        Optional<Usuario> usuario = usuarioRepository.findByEmailAndSenha(email, senha);
+        Usuario usuario = usuarioRepository.findByEmailAndSenha(email, senha)
+                .orElseThrow(() -> new RegraDeNegocioException("Email ou senha inválidos."));
 
-        if (usuario.isPresent()) {
-            return usuario.get().toDTO();
-        }
-
-        return null;
+        return usuario.toDTO();
     }
 
 }
